+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 1. [](#bugfix)
     * Regression: Fixed fatal error in blueprints [#2811](https://github.com/getgrav/grav/issues/2811)
     * Regression: Fixed bad method call in FlexDirectory::getAuthorizeRule()
+    * Regression: Fixed fatal error in admin if the site has custom permissions in `onAdminRegisterPermissions`
 
 # v1.7.0-rc.5
 ## 02/03/2020